refactor code
The Home.module.css file was deleted as a part of this commit. Additionally, several sections in the codebase were refactored including Animation.tsx being renamed to CameraAnimation.tsx, integrating new logic, and changes in Island component. Further, the Sky and Island components were moved to a new directory, a new Progress component was created, and enhancements were made to improve the performance and readability of the